% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_server.R
\name{dataServer}
\alias{dataServer}
\title{Shiny module for loading data - server side}
\usage{
dataServer(id)
}
\arguments{
\item{id}{character, id of module}
}
\value{
reactive object, loaded data
}
\description{
Shiny module for loading data - server side
}
