% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ui.R
\name{dataUI}
\alias{dataUI}
\title{Shiny module for loading data - UI side}
\usage{
dataUI(id)
}
\arguments{
\item{id}{character, id of module}
}
\value{
UI object
}
\description{
Shiny module for loading data - UI side
}
