% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_server.R
\name{fitServer}
\alias{fitServer}
\title{Shiny module for model fitting - server side}
\usage{
fitServer(
  id,
  loadedData,
  message,
  selectedResponse,
  selectedResponseDosage,
  inFile
)
}
\arguments{
\item{id}{character. id of module}

\item{loadedData}{data frame, data after subsetting}

\item{message}{list, messages from data loading tab for report}

\item{selectedResponse}{character, selected responses}

\item{selectedResponseDosage}{character, selected dosage}

\item{inFile}{reactive, input file}
}
\value{
list with fitted models and analysis data
}
\description{
Shiny module for model fitting - server side
}
