% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fitUI}
\alias{fitUI}
\title{Shiny module for model fitting - UI side}
\usage{
fitUI(id)
}
\arguments{
\item{id}{character, id of module}
}
\value{
UI object
}
\description{
Shiny module for model fitting - UI side
}
