% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{globalFunction}
\alias{globalFunction}
\title{Global function - run before serverFunction and uiFunction}
\usage{
globalFunction()
}
\value{
no return value
}
\description{
Global function - run before serverFunction and uiFunction
}
