% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{readDataBMD}
\alias{readDataBMD}
\title{Read BMD data for analysis}
\usage{
readDataBMD(fileName, filePath, separator, decimal)
}
\arguments{
\item{fileName}{character, name of the file (without full path)}

\item{filePath}{path name, where is the .txt or .csv file located}

\item{separator}{delimiter for the dataset}

\item{decimal}{character, used for decimal points}
}
\value{
dataframe
}
\description{
Updated because of decimal separator, extra argument compared to f.scan()
}
