% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_server.R
\name{resultsServer}
\alias{resultsServer}
\title{Shiny module for displaying results - server side}
\usage{
resultsServer(id, analysis_result)
}
\arguments{
\item{id}{character. id of module}

\item{fit_result}{analysis results}
}
\value{
list with fitted models and analysis data
}
\description{
Shiny module for displaying results - server side
}
