% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_ui.R
\name{resultsUI}
\alias{resultsUI}
\title{Shiny module for the results - UI side}
\usage{
resultsUI(id)
}
\arguments{
\item{id}{character, id of module}
}
\value{
UI object
}
\description{
Shiny module for the results - UI side
}
