% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Run the shiny application}
\usage{
runShiny(installDependencies = FALSE, ...)
}
\arguments{
\item{installDependencies}{logical, whether to install all dependent packages or not}

\item{...}{further arguments that can be passed to \code{\link[shiny]{runApp}}}
}
\value{
no return value
}
\description{
Run the shiny application
}
