% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Main server function}
\usage{
server(input, output)
}
\arguments{
\item{input}{Shiny object}

\item{output}{Shiny object}
}
\value{
no return value
}
\description{
Main server function
}
