% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{serverFunction}
\alias{serverFunction}
\title{Main server function}
\usage{
serverFunction(input, output, session)
}
\arguments{
\item{input}{Shiny object}

\item{output}{Shiny object}

\item{session}{Shiny object}
}
\value{
no return value
}
\description{
Main server function
}
