% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{fitServer}
\alias{fitServer}
\title{Shiny module for model fitting - server side}
\usage{
fitServer(id, loadedData, message, dataType, isLitter, selectedResponse)
}
\arguments{
\item{id}{character. id of module}

\item{loadedData}{data frame, data after subsetting}

\item{message}{list, messages from data loading tab for report}

\item{dataType}{character, type of loaded data (e.g. continuous summary or quantal)}

\item{isLitter}{logical}

\item{selectedResponse}{character, selected responses}
}
\value{
list with fitted models and analysis data
}
\description{
Shiny module for model fitting - server side
}
\author{
mvarewyck
}
